IPV4 implemented for test

+++ checksum.cpp
@@ -1,16 +1,16 @@
 #include "checksum.h"
- 
+
 /**
  * @brief Checks a 16bits Integer check sum w/ one's complement
  * @param *data is a pointer to the prealocated data
  * @param dataLenght is the lenght of data
  * @param checkSumByte indicates where the check sum byte is in the data
  */                                 
-int check_csum(int *data, unsigned int dataLenght, unsigned int checkSumByte)
+int16_t check_csum(int16_t *data, uint16_t dataLenght, uint16_t checkSumByte)
 {
     checkSumByte--;
-    int sum = 0;
-    int *data_start = data;     // keeps the pointer reference
+    int16_t sum = 0;
+    int16_t *data_start = data;     // keeps the pointer reference
     for(; dataLenght; dataLenght--){
         // ignores the checkSumByte for computing the sum
         if(dataLenght != checkSumByte){
@@ -29,12 +29,12 @@ int check_csum(int *data, unsigned int dataLenght, unsigned int checkSumByte)
  * @param *data is a pointer to the prealocated data
  * @param dataLenght is the lenght of data
  * @param checkSumByte indicates where the check sum byte is in the data
- */                                
-int make_csum(int *data, unsigned int dataLenght, unsigned int checkSumByte)   
+ */                                 
+int16_t make_csum(int16_t *data, uint16_t dataLenght, uint16_t checkSumByte)
 {
     checkSumByte--;
-    int sum = 0;
-    int *data_start = data;     // keeps the pointer reference
+    int16_t sum = 0;
+    int16_t *data_start = data;     // keeps the pointer reference
     for(; dataLenght; dataLenght--){
         sum += *data++;
         if(sum & 0xFFFF0000) sum = (sum & 0xFFFF) +1; // ones' complement
